+++ custom_components/solarcharger/chargers/sc_option_state.py
@@ -44,7 +44,7 @@ from ..const import (
     SENSOR_CONSUMED_ENERGY_TODAY,
     SENSOR_CONSUMED_POWER,
     SENSOR_DELTA_ALLOCATED_POWER,
-    SENSOR_SELF_DERATED_TODAY,
+    SENSOR_SELF_DEPOWER_TODAY,
     SENSOR_SHARE_ALLOCATION,
     SENSOR_SYNC_UPDATE,
     SWITCH,
@@ -124,10 +124,10 @@ class ScOptionState(ScConfigState):
         )
 
     @cached_property
-    def self_derated_today_entity_id(self) -> str:
-        """Return the self-derated today entity ID."""
+    def self_depower_today_entity_id(self) -> str:
+        """Return the self-depower today entity ID."""
         return compose_entity_id(
-            SENSOR, self._subentry.unique_id, SENSOR_SELF_DERATED_TODAY
+            SENSOR, self._subentry.unique_id, SENSOR_SELF_DEPOWER_TODAY
         )
 
     @cached_property
@@ -886,10 +886,10 @@ class ScOptionState(ScConfigState):
 
     # ----------------------------------------------------------------------------
     # Need entity to keep state across reboots.
-    def get_self_derated_today(self) -> int:
-        """Get self-derated today."""
+    def get_self_depower_today(self) -> int:
+        """Get self-depower today."""
 
-        return self.get_integer(self.self_derated_today_entity_id)
+        return self.get_integer(self.self_depower_today_entity_id)
 
     # ----------------------------------------------------------------------------
     def is_fast_charge_mode(self) -> bool: